jobfind_ai/requirements.txt
streamlit==1.28.1
openai==0.28.1
python-dotenv==1.0.0 